requirements.txt
streamlit
requests
python-dotenv
beautifulsoup4
lxml
